Extract EndCall
