+++ packages/react-composites/src/composites/CallComposite/components/buttons/EndCall.tsx
@@ -1,0 +1,34 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import { concatStyleSets } from '@fluentui/react';
+import { ControlBarButtonStyles, EndCallButton } from '@internal/react-components';
+import React, { useMemo } from 'react';
+import { usePropsFor } from '../../hooks/usePropsFor';
+import { CallControlDisplayType } from '../../types/CallControlOptions';
+import { groupCallLeaveButtonCompressedStyle, groupCallLeaveButtonStyle } from './styles';
+
+/** @private */
+export const EndCall = (props: {
+  displayType?: CallControlDisplayType;
+  styles?: ControlBarButtonStyles;
+}): JSX.Element => {
+  const compactMode = props.displayType === 'compact';
+  const hangUpButtonProps = usePropsFor(EndCallButton);
+  const styles = useMemo(
+    () =>
+      concatStyleSets(
+        compactMode ? groupCallLeaveButtonCompressedStyle : groupCallLeaveButtonStyle,
+        props.styles ?? {}
+      ),
+    [props.styles]
+  );
+  return (
+    <EndCallButton
+      data-ui-id="call-composite-hangup-button"
+      {...hangUpButtonProps}
+      styles={styles}
+      showLabel={!compactMode}
+    />
+  );
+};
